move classes to correct package (instead of "iternal")
and fix lots of warnings
