+++ hibernate-reactive-core/src/main/java/org/hibernate/reactive/query/sqm/internal/AggregatedSelectReactiveQueryPlan.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  * Copyright: Red Hat Inc. and Hibernate Authors
  */
-package org.hibernate.reactive.query.sqm.iternal;
+package org.hibernate.reactive.query.sqm.internal;
 
 import java.util.ArrayList;
 import java.util.List;
