+++ hibernate-reactive-core/src/main/java/org/hibernate/reactive/query/sqm/internal/ReactiveMultiTableDeleteQueryPlan.java
@@ -3,37 +3,37 @@
  * SPDX-License-Identifier: Apache-2.0
  * Copyright: Red Hat Inc. and Hibernate Authors
  */
-package org.hibernate.reactive.query.sqm.iternal;
+package org.hibernate.reactive.query.sqm.internal;
 
 import java.util.concurrent.CompletionStage;
 
 import org.hibernate.action.internal.BulkOperationCleanupAction;
 import org.hibernate.query.spi.DomainQueryExecutionContext;
 import org.hibernate.query.sqm.internal.DomainParameterXref;
-import org.hibernate.query.sqm.tree.update.SqmUpdateStatement;
+import org.hibernate.query.sqm.tree.delete.SqmDeleteStatement;
 import org.hibernate.reactive.query.sql.spi.ReactiveNonSelectQueryPlan;
 import org.hibernate.reactive.query.sqm.mutation.spi.ReactiveSqmMultiTableMutationStrategy;
 
 /**
- * @see org.hibernate.query.sqm.internal.MultiTableUpdateQueryPlan
+ * @see org.hibernate.query.sqm.internal.MultiTableDeleteQueryPlan
  */
-public class ReactiveMultiTableUpdateQueryPlan implements ReactiveNonSelectQueryPlan {
-	private final SqmUpdateStatement sqmUpdate;
+public class ReactiveMultiTableDeleteQueryPlan implements ReactiveNonSelectQueryPlan {
+	private final SqmDeleteStatement<?> sqmDelete;
 	private final DomainParameterXref domainParameterXref;
-	private final ReactiveSqmMultiTableMutationStrategy mutationStrategy;
+	private final ReactiveSqmMultiTableMutationStrategy deleteStrategy;
 
-	public ReactiveMultiTableUpdateQueryPlan(
-			SqmUpdateStatement sqmUpdate,
+	public ReactiveMultiTableDeleteQueryPlan(
+			SqmDeleteStatement<?> sqmDelete,
 			DomainParameterXref domainParameterXref,
-			ReactiveSqmMultiTableMutationStrategy mutationStrategy) {
-		this.sqmUpdate = sqmUpdate;
+			ReactiveSqmMultiTableMutationStrategy deleteStrategy) {
+		this.sqmDelete = sqmDelete;
 		this.domainParameterXref = domainParameterXref;
-		this.mutationStrategy = mutationStrategy;
+		this.deleteStrategy = deleteStrategy;
 	}
 
 	@Override
 	public CompletionStage<Integer> executeReactiveUpdate(DomainQueryExecutionContext executionContext) {
-		BulkOperationCleanupAction.schedule( executionContext.getSession(), sqmUpdate );
-		return mutationStrategy.reactiveExecuteUpdate( sqmUpdate, domainParameterXref, executionContext );
+		BulkOperationCleanupAction.schedule( executionContext.getSession(), sqmDelete );
+		return deleteStrategy.reactiveExecuteDelete( sqmDelete, domainParameterXref, executionContext );
 	}
 }
