+++ hibernate-reactive-core/src/main/java/org/hibernate/reactive/query/sqm/internal/ReactiveMultiTableUpdateQueryPlan.java
@@ -3,37 +3,37 @@
  * SPDX-License-Identifier: Apache-2.0
  * Copyright: Red Hat Inc. and Hibernate Authors
  */
-package org.hibernate.reactive.query.sqm.iternal;
+package org.hibernate.reactive.query.sqm.internal;
 
 import java.util.concurrent.CompletionStage;
 
 import org.hibernate.action.internal.BulkOperationCleanupAction;
 import org.hibernate.query.spi.DomainQueryExecutionContext;
 import org.hibernate.query.sqm.internal.DomainParameterXref;
-import org.hibernate.query.sqm.tree.delete.SqmDeleteStatement;
+import org.hibernate.query.sqm.tree.update.SqmUpdateStatement;
 import org.hibernate.reactive.query.sql.spi.ReactiveNonSelectQueryPlan;
 import org.hibernate.reactive.query.sqm.mutation.spi.ReactiveSqmMultiTableMutationStrategy;
 
 /**
- * @see org.hibernate.query.sqm.internal.MultiTableDeleteQueryPlan
+ * @see org.hibernate.query.sqm.internal.MultiTableUpdateQueryPlan
  */
-public class ReactiveMultiTableDeleteQueryPlan implements ReactiveNonSelectQueryPlan {
-	private final SqmDeleteStatement sqmDelete;
+public class ReactiveMultiTableUpdateQueryPlan implements ReactiveNonSelectQueryPlan {
+	private final SqmUpdateStatement<?> sqmUpdate;
 	private final DomainParameterXref domainParameterXref;
-	private final ReactiveSqmMultiTableMutationStrategy deleteStrategy;
+	private final ReactiveSqmMultiTableMutationStrategy mutationStrategy;
 
-	public ReactiveMultiTableDeleteQueryPlan(
-			SqmDeleteStatement sqmDelete,
+	public ReactiveMultiTableUpdateQueryPlan(
+			SqmUpdateStatement<?> sqmUpdate,
 			DomainParameterXref domainParameterXref,
-			ReactiveSqmMultiTableMutationStrategy deleteStrategy) {
-		this.sqmDelete = sqmDelete;
+			ReactiveSqmMultiTableMutationStrategy mutationStrategy) {
+		this.sqmUpdate = sqmUpdate;
 		this.domainParameterXref = domainParameterXref;
-		this.deleteStrategy = deleteStrategy;
+		this.mutationStrategy = mutationStrategy;
 	}
 
 	@Override
 	public CompletionStage<Integer> executeReactiveUpdate(DomainQueryExecutionContext executionContext) {
-		BulkOperationCleanupAction.schedule( executionContext.getSession(), sqmDelete );
-		return deleteStrategy.reactiveExecuteDelete( sqmDelete, domainParameterXref, executionContext );
+		BulkOperationCleanupAction.schedule( executionContext.getSession(), sqmUpdate );
+		return mutationStrategy.reactiveExecuteUpdate( sqmUpdate, domainParameterXref, executionContext );
 	}
 }
