+++ hibernate-reactive-core/src/main/java/org/hibernate/reactive/query/sqm/mutation/internal/temptable/ReactiveExecuteWithTemporaryTableHelper.java
@@ -71,7 +71,6 @@ public final class ReactiveExecuteWithTemporaryTableHelper {
 			Function<SharedSessionContractImplementor, String> sessionUidAccess,
 			JdbcParameterBindings jdbcParameterBindings,
 			ExecutionContext executionContext) {
-		final SessionFactoryImplementor factory = executionContext.getSession().getFactory();
 
 		final TableGroup mutatingTableGroup = sqmConverter.getMutatingTableGroup();
