+++ hibernate-reactive-core/src/main/java/org/hibernate/reactive/query/sqm/mutation/internal/temptable/ReactiveGlobalTemporaryTableInsertStrategy.java
@@ -23,9 +23,9 @@ import org.hibernate.reactive.query.sqm.mutation.spi.ReactiveSqmMultiTableInsert
 public class ReactiveGlobalTemporaryTableInsertStrategy extends GlobalTemporaryTableStrategy
 		implements ReactiveGlobalTemporaryTableStrategy, ReactiveSqmMultiTableInsertStrategy {
 
-	private final CompletableFuture<Void> tableCreatedStage = new CompletableFuture();
+	private final CompletableFuture<Void> tableCreatedStage = new CompletableFuture<>();
 
-	private final CompletableFuture<Void> tableDroppedStage = new CompletableFuture();
+	private final CompletableFuture<Void> tableDroppedStage = new CompletableFuture<>();
 
 	private boolean prepared;
